debian installer: disable automatic partitioning
Systems use a wide variety of disk devices (sd, nvme, emmc), and so
automatic partitioning frequently kills the usb install media when the
target disk is not /dev/sda. Installing to a different disk type has
required building a new ISO and writing new install media.

This removes the automatic disk partitioning stuff, which means that
partitioning is now a manual step in the install process.