fetch_sources: always update mesa_ci to master before executing
In some cases, changes to mesa_ci are not backward-compatible.  In
these situations, fetch_sources will fail when using an older mesa_ci
residing on a build system.

It should always be safe to use master for mesa_ci, when running
fetch_sources.  If a developer wishes to run with an older mesa_ci,
that revision can be specified in the parameters to fetch_sources.