piglit-test: limit # of piglit jobs on simdrm based on RAM size
Some systems have a lot of logical CPUs but not enough memory to prevent
OOM events, so this limits the number of concurrent piglit jobs based on
the system RAM size. The ratio used (3/8) is from 'stable' servers
running piglit on simdrm, which are already running with 50% of logical
CPUs utilized.

Based on some light testing, this seems to allow piglit to get through a
run with simdrm on a memory-limited server with no OOM.