fetch_sources: do not consider platform support when fetching
DependencyGraph drops projects that cannot be executed on the current
version due to hardware support.  This prevents unnecessary scheduling
of components.

DependencyGraph is also used to build a list of source projects that
must be fetched.  It may be that the version on disk prior to fetching
does not support the target components.  After fetching, the version
will be updated.  For this use case, DependencyGraph must not consider
platform support.