vulkancts: handle new vk cts mustpass layout
The new layout is a single file that includes a list of mustpass files
to use. This change adds support for that, as well as preserves support
for a single mustpass that lists all tests (used by current vk cts
releases.)

This removes support for the versioned mustpass, since there
should be nothing in CI that uses a vk cts old enough to need that.